# docs/swagger.yaml
definitions:
  models.Menu:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      isAvailable:
        type: boolean
      name:
        type: string
      price:
        format: float64
        type: number
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      updatedAt:
        type: string
    type: object
  models.MenuUpdateInput:
    properties:
      description:
        type: string
      isAvailable:
        type: boolean
      name:
        type: string
      price:
        type: number
      productsIDs:
        items:
          type: integer
        type: array
    type: object
  models.Order:
    properties:
      createdAt:
        type: string
      deliveredAt:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      preparedAt:
        type: string
      status:
        $ref: '#/definitions/models.OrderStatus'
      ticketNumber:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    required:
    - status
    - user
    type: object
  models.OrderItem:
    properties:
      id:
        type: integer
      orderContentDescription:
        type: string
      orderContentImage:
        type: string
      orderContentName:
        type: string
      orderContentPrice:
        format: float64
        type: number
      orderID:
        type: integer
      quantity:
        type: integer
    type: object
  models.OrderStatus:
    enum:
    - created
    - inPreparation
    - prepared
    - delivered
    type: string
    x-enum-varnames:
    - Created
    - InPreparation
    - Prepared
    - Delivered
  models.OrderUpdateInput:
    properties:
      items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      ticketNumber:
        type: string
    type: object
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.ProductCategory'
      categoryID:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      isAvailable:
        type: boolean
      name:
        type: string
      price:
        format: float64
        type: number
      updatedAt:
        type: string
    type: object
  models.ProductCategory:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        minLength: 8
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      updatedAt:
        type: string
    required:
    - email
    - password
    type: object
...
  models.UserRole:
    enum:
    - admin
    - greeter
    - order_picker
    - manager
    type: string
    x-enum-varnames:
    - Admin
    - Greeter
    - OrderPicker
    - Manager
...
info:
  contact: {}
  description: Order management application for Wacdo
  title: Wacdo
  version: "1.0"
paths:
  /menu:
    post:
      consumes:
      - application/json
      description: Créer un nouveau menu
      parameters:
      - description: Données du menu
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: Données invalides
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Menus
  /menus:
    get:
      description: Récupérer tous les menus
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Menus
  /menus/{id}:
    delete:
      description: Supprimer un menu
      parameters:
      - description: ID du menu
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Message de succès
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID invalide
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Produit non trouvé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Menus
    get:
      description: Récupérer un menu par son ID
      parameters:
      - description: ID du menu
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: ID invalide
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Menu non trouvé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Mettre à jour un menu existant
      parameters:
      - description: ID du menu
        in: path
        name: id
        required: true
        type: integer
      - description: Données de mise à jour
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.MenuUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Menu'
        "400":
          description: Données invalides
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Produit non trouvé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Menus
  /order:
    post:
      consumes:
      - application/json
      description: Créer une nouvelle commande
      parameters:
      - description: Données de la commande
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Données invalides
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Orders
  /orders:
    get:
      description: Récupérer toutes les commandes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
      security:
      - BearerAuth: []
      tags:
      - Orders
  /orders/{id}:
    get:
      description: Récupérerer une commande par son ID
      parameters:
      - description: ID de la commande
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: ID invalide
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Commande non trouvée
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Mettre à jour une commande existante
      parameters:
      - description: ID de la commande
        in: path
        name: id
        required: true
        type: integer
      - description: Données de mise à jour
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.OrderUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Données invalides
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Commande non trouvée
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Orders
...
  /orders/{id}/in-preparation:
    patch:
      consumes:
      - application/json
      description: Indiquer que la commande est en préparation
      parameters:
      - description: ID de la commande
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Données invalides
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Commande non trouvée
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Orders
  /orders/{id}/prepared:
    patch:
      consumes:
      - application/json
      description: Indiquer que la commande a été préparée
      parameters:
      - description: ID de la commande
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: Données invalides
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Commande non trouvée
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur interne
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"