# models/order_status_model.go
package models

type OrderStatus string

const (
	Created       OrderStatus = "created"
	InPreparation OrderStatus = "inPreparation"
	Prepared      OrderStatus = "prepared"
	Delivered     OrderStatus = "delivered"
)